move all under src dir
