+++ package.json
@@ -14,11 +14,11 @@
   "dependencies": {
     "@tanstack/react-query": "^4.20.4",
     "@tanstack/react-query-devtools": "^4.20.4",
-    "@types/node": "18.11.14",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.9",
-    "eslint": "8.29.0",
-    "eslint-config-next": "13.0.6",
+    "@types/node": "18.15.11",
+    "@types/react": "18.0.33",
+    "@types/react-dom": "18.0.11",
+    "eslint": "8.38.0",
+    "eslint-config-next": "13.3.0",
     "next": "^13.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
